+++ test/snapshot/core/closures_capture/intended/main.go
@@ -6,7 +6,7 @@ func main() {
 	factor := 3
 	_ = factor
 	mul := func(v int) int {
-		return (v * factor)
+		return int(int32((int32(v) * int32(factor))))
 	}
 	_ = mul
 	factor = 4
