+++ test/snapshot/core/default_args/intended/main.go
@@ -3,12 +3,12 @@ package main
 import "snapshot/hxrt"
 
 func add(a int, b int) int {
-	return (a + b)
+	return int(int32((int32(a) + int32(b))))
 }
 
 func main() {
 	local := func(v int) int {
-		return (v + 1)
+		return int(int32((int32(v) + int32(1))))
 	}
 	_ = local
 	hxrt.Println(add(1, 2))
