+++ test/snapshot/go_native/goroutine_smoke/intended/main.go
@@ -40,7 +40,7 @@ func (self *go___Chan) recv() any {
 	}
 	var value any = self.queue[self.readIndex]
 	_ = value
-	self.readIndex = (self.readIndex + 1)
+	self.readIndex = int(int32((self.readIndex + 1)))
 	return value
 }
